add authorization middleware in routes & models

+++ src/db/models/parkingSpot.js
@@ -2,18 +2,28 @@ const database = require("../config/mongodb");
 const { ObjectId } = require("mongodb");
 
 class ParkingSpotModels {
-  static async getAll() {
+  static async getAll({ role, authorId }) {
+    let opt = {};
+    if (role === "vendor") {
+      opt = {
+        authorId,
+      };
+    }
+
     const parkSpots = await database
       .collection("parkingSpots")
-      .find()
+      .find(opt)
       .toArray();
     return parkSpots;
   }
 
-  static async getById(id) {
-    const parkingSpot = await database
-      .collection("parkingSpots")
-      .findOne({ _id: new ObjectId(String(id)) });
+  static async getById({ id, authorId, role }) {
+    let opt = { _id: new ObjectId(String(id)) };
+    if (role === "vendor") {
+      opt["$and"] = [{ authorId: new ObjectId(String(authorId)) }];
+    }
+
+    const parkingSpot = await database.collection("parkingSpots").findOne(opt);
     if (!parkingSpot) throw { name: "ParkingSpotNotFound" };
     return parkingSpot;
   }
@@ -26,7 +36,11 @@ class ParkingSpotModels {
     carSpot,
     motorFee,
     carFee,
+    authorId,
+    role,
   }) {
+    if (role !== "vendor")
+      throw { name: "Your account unauthorized to create parking spot" };
     const result = await database.collection("parkingSpots").insertOne({
       name,
       address,
@@ -35,11 +49,11 @@ class ParkingSpotModels {
       carSpot,
       motorFee,
       carFee,
-      authorId: new ObjectId(String("66d6d3d0cf201705437e09cc")),
+      authorId: new ObjectId(String(authorId)),
     });
     return "Success create parking spot";
   }
-  
+
   static async updateParkingSpot({
     id,
     name,
@@ -49,7 +63,10 @@ class ParkingSpotModels {
     carSpot,
     motorFee,
     carFee,
+    role,
   }) {
+    if (role !== "vendor")
+      throw { name: "Your account unauthorized to update parking spot data" };
     const result = await database.collection("parkingSpots").updateOne(
       {
         _id: new ObjectId(String(id)),
@@ -59,13 +76,15 @@ class ParkingSpotModels {
     return "Success update parking spot";
   }
 
-
-  static async deleteParkingSpot(id) {
-    const result = await database.collection("parkingSpots").deleteOne(
-      {
-        _id: new ObjectId(String(id)),
-      }
-    )
+  static async deleteParkingSpot({ id, authorId, role }) {
+    if (role !== "vendor")
+      throw { name: "Your account unauthorized to delete parking spot" };
+    const result = await database.collection("parkingSpots").deleteOne({
+      $and: [
+        { _id: new ObjectId(String(id)) },
+        { authorId: new ObjectId(String(authorId)) },
+      ],
+    });
     return { result: "Success delete parking spot" };
   }
 }
